combined log: enclose remote_user in dquotes
The request, http_referer and http_user_agent are all enclosed in
dquotes as they may contain whitespace. To be consistent with this, the
remote_user should also be enclosed in dquotes. The additional benefits
of doing so are the simplified logic for log line parsing and easier
grokking for tired eyes.

Example:

```
127.0.0.1 - user with spaces [04/Apr/2025:11:09:30 +0100] ...
```

becomes

```
127.0.0.1 - "user with spaces" [04/Apr/2025:11:09:30 +0100] ...
```